requirements.txt
google-generativeai
awan-llm-api
pypdf
